Compatibility with Cisco Catalyst Center version 3.1.6.0

+++ plugins/action/tag_info.py
@@ -2,8 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # Copyright (c) 2021, Cisco Systems
-# GNU General Public License v3.0+ (see LICENSE or
-# https://www.gnu.org/licenses/gpl-3.0.txt)
+# GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
 
@@ -33,8 +32,8 @@ argument_spec.update(
         additionalInfo_nameSpace=dict(type="str"),
         additionalInfo_attributes=dict(type="str"),
         level=dict(type="str"),
-        offset=dict(type="float"),
-        limit=dict(type="float"),
+        offset=dict(type="int"),
+        limit=dict(type="int"),
         size=dict(type="str"),
         field=dict(type="str"),
         sortBy=dict(type="str"),
